requirements.txt
# Data & Analysis
pandas==2.3.3
numpy==2.3.4
matplotlib==3.10.7
plotly==6.3.1

# Visualization helpers
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
kiwisolver==1.4.9
packaging==25.0
pillow==12.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
six==1.17.0
narwhals==2.10.1

# Database
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10

# Environment / Config
python-dotenv==1.0.1